+++ src/main/java/com/allclear/tastytrack/domain/review/repository/ReviewRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.allclear.tastytrack.domain.review.entity.Review;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
-	List<Review> findAllByRestaurantIdOrderByCreatedAtDesc(Long restaurantId);
+	List<Review> findAllByRestaurantIdOrderByCreatedAtDesc(Integer restaurantId);
 }
